+++ source/alt_decoder.c
@@ -1137,7 +1137,7 @@ __user__decoder_int_thread (DECO * calc)
 }
 
 void
-matface (int *par, int *nin, double *in_r, double *in_i,
+matface (int *par, int *nin, short *in_r, short *in_i,
 	 int *nout, double *out_r, double *out_i, double *upar)
 {
   ulong nb = 7 + par[0] * par[6];
